[move-ide] Symbolication refactoring
