+++ external-crates/move/crates/move-analyzer/src/analysis/parsing_analysis.rs
@@ -3,10 +3,12 @@
 
 use crate::{
     symbols::{
-        AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, CursorContext,
-        CursorDefinition, CursorPosition, DefMap, ModuleDefs, References, UseDef, UseDefMap,
         add_member_use_def, ignored_function, parsed_address,
         parsing_leading_and_mod_names_to_map_key, parsing_mod_def_to_map_key,
+        types::{
+            AutoImportInsertionInfo, AutoImportInsertionKind, CallInfo, CursorContext,
+            CursorDefinition, CursorPosition, DefMap, ModuleDefs, References, UseDef, UseDefMap,
+        },
     },
     utils::{loc_end_to_lsp_position_opt, loc_start_to_lsp_position_opt},
 };
