+++ external-crates/move/crates/move-analyzer/src/analysis/typing_analysis.rs
@@ -4,8 +4,10 @@
 use crate::{
     compiler_info::CompilerInfo,
     symbols::{
-        DefInfo, DefMap, LocalDef, MemberDefInfo, ModuleDefs, References, UseDef, UseDefMap,
         add_member_use_def, expansion_mod_ident_to_map_key, find_datatype, type_def_loc,
+        types::{
+            DefInfo, DefMap, LocalDef, MemberDefInfo, ModuleDefs, References, UseDef, UseDefMap,
+        },
     },
     utils::{ignored_function, loc_start_to_lsp_position_opt},
 };
