+++ external-crates/move/crates/move-analyzer/src/completions/dot.rs
@@ -5,7 +5,10 @@
 
 use crate::{
     completions::utils::{call_completion_item, mod_defs},
-    symbols::{DefInfo, FunType, Symbols, type_to_ide_string},
+    symbols::{
+        ide_strings::type_to_ide_string,
+        types::{DefInfo, FunType, Symbols},
+    },
     utils::lsp_position_to_loc,
 };
 use lsp_types::{
