+++ external-crates/move/crates/move-analyzer/src/completions/mod.rs
@@ -10,7 +10,10 @@ use crate::{
         utils::{PRIMITIVE_TYPE_COMPLETIONS, completion_item},
     },
     context::Context,
-    symbols::{self, CursorContext, PrecomputedPkgInfo, SymbolicatorRunner, Symbols},
+    symbols::{
+        self, SymbolicatorRunner,
+        types::{CursorContext, PrecomputedPkgInfo, Symbols},
+    },
 };
 use lsp_server::{Message, Request, Response};
 use lsp_types::{CompletionItem, CompletionItemKind, CompletionParams, Position};
