+++ external-crates/move/crates/move-analyzer/src/completions/snippets.rs
@@ -6,7 +6,7 @@
 
 use crate::{
     completions::utils::mod_defs,
-    symbols::{CursorContext, CursorDefinition, DefInfo, Symbols},
+    symbols::types::{CursorContext, CursorDefinition, DefInfo, Symbols},
 };
 use lsp_types::{CompletionItem, CompletionItemKind, Documentation, InsertTextFormat, Position};
 use move_command_line_common::files::FileHash;
