+++ external-crates/move/crates/move-analyzer/src/completions/utils.rs
@@ -4,10 +4,15 @@
 use std::path::PathBuf;
 
 use crate::symbols::{
-    AutoImportInsertionInfo, AutoImportInsertionKind, CompiledPkgInfo, CursorContext, DefInfo,
-    ModuleDefs, Symbols, SymbolsComputationData, compute_symbols_parsed_program,
-    compute_symbols_pre_process, mod_ident_to_ide_string, ret_type_to_ide_str,
-    type_args_to_ide_string, type_list_to_ide_string,
+    compute_symbols_parsed_program, compute_symbols_pre_process,
+    ide_strings::{
+        mod_ident_to_ide_string, ret_type_to_ide_str, type_args_to_ide_string,
+        type_list_to_ide_string,
+    },
+    types::{
+        AutoImportInsertionInfo, AutoImportInsertionKind, CompiledPkgInfo, CursorContext, DefInfo,
+        ModuleDefs, Symbols, SymbolsComputationData,
+    },
 };
 use lsp_types::{
     CompletionItem, CompletionItemKind, CompletionItemLabelDetails, InsertTextFormat, Position,
