+++ external-crates/move/crates/move-analyzer/src/symbols/mod.rs
@@ -56,6 +56,15 @@ use crate::{
     compiler_info::CompilerInfo,
     context::Context,
     diagnostics::{lsp_diagnostics, lsp_empty_diagnostics},
+    symbols::{
+        ide_strings::{const_val_to_ide_string, mod_ident_to_ide_string},
+        types::{
+            AnalyzedPkgInfo, CompiledPkgInfo, CompiledProgram, CursorContext, DefInfo, DefMap,
+            FieldDef, FieldOrderInfo, FileModules, FileUseDefs, FunType, MemberDef, MemberDefInfo,
+            ModuleDefs, PrecomputedPkgInfo, References, Symbols, SymbolsComputationData, UseDef,
+            UseDefMap, VariantInfo,
+        },
+    },
     utils::{loc_start_to_lsp_position_opt, lsp_position_to_loc},
 };
 
@@ -70,9 +79,7 @@ use lsp_types::{
 };
 use sha2::{Digest, Sha256};
 use std::{
-    cmp,
     collections::{BTreeMap, BTreeSet},
-    fmt,
     path::{Path, PathBuf},
     sync::{Arc, Condvar, Mutex},
     thread,
@@ -89,31 +96,18 @@ use vfs::{
 use move_command_line_common::files::FileHash;
 use move_compiler::{
     PASS_CFGIR, PASS_PARSER, PASS_TYPING,
-    command_line::compiler::{FullyCompiledProgram, construct_pre_compiled_lib},
+    command_line::compiler::construct_pre_compiled_lib,
     editions::{Edition, FeatureGate, Flavor},
-    expansion::{
-        ast::{self as E, AbilitySet, ModuleIdent, ModuleIdent_, Value, Value_, Visibility},
-        name_validation::{
-            IMPLICIT_STD_MEMBERS, IMPLICIT_STD_MODULES, IMPLICIT_SUI_MEMBERS, IMPLICIT_SUI_MODULES,
-            ModuleMemberKind,
-        },
-    },
+    expansion::ast::{self as E, ModuleIdent, ModuleIdent_, Visibility},
     linters::LintLevel,
-    naming::ast::{
-        DatatypeTypeParameter, Neighbor, StructFields, Type, Type_, TypeName_, VariantFields,
-    },
+    naming::ast::{DatatypeTypeParameter, StructFields, Type, Type_, TypeName_, VariantFields},
     parser::ast::{self as P, DocComment},
     shared::{
-        Identifier, Name, NamedAddressMap, NamedAddressMaps, files::MappedFiles,
-        unique_map::UniqueMap,
+        Identifier, NamedAddressMap, NamedAddressMaps, files::MappedFiles, unique_map::UniqueMap,
     },
-    typing::{
-        ast::{Exp, ExpListItem, ModuleDefinition, SequenceItem, SequenceItem_, UnannotatedExp_},
-        visitor::TypingVisitorContext,
-    },
+    typing::{ast::ModuleDefinition, visitor::TypingVisitorContext},
     unit_test::filter_test_members::UNIT_TEST_POISON_FUN_NAME,
 };
-use move_core_types::{account_address::AccountAddress, parsing::address::NumericalAddress};
 use move_ir_types::location::*;
 use move_package::{
     compilation::{build_plan::BuildPlan, compiled_package::ModuleFormat},
@@ -122,601 +116,13 @@ use move_package::{
 };
 use move_symbol_pool::Symbol;
 
+pub mod ide_strings;
+pub mod types;
+
 const MANIFEST_FILE_NAME: &str = "Move.toml";
-const STD_LIB_PKG_ADDRESS: &str = "0x1";
-const SUI_LIB_PKG_ADDRESS: &str = "0x2";
-
-/// Information about compiled program (ASTs at different levels)
-#[derive(Clone)]
-pub struct CompiledProgram {
-    pub parsed: P::Program,
-    pub typed_modules: UniqueMap<ModuleIdent, ModuleDefinition>,
-}
-
-/// Package data used during compilation and analysis
-#[derive(Clone)]
-struct AnalyzedPkgInfo {
-    /// Cached fully compiled program representing dependencies
-    program_deps: Arc<FullyCompiledProgram>,
-    /// Cached symbols computation data for dependencies
-    symbols_data: Option<Arc<SymbolsComputationData>>,
-    /// Compiled user program
-    program: Option<Arc<CompiledProgram>>,
-    /// Mapping from file paths to file hashes
-    file_hashes: Arc<BTreeMap<PathBuf, FileHash>>,
-}
-
-/// Information about the compiled package and data structures
-/// computed during compilation and analysis
-#[derive(Clone)]
-pub struct CompiledPkgInfo {
-    /// Package path
-    path: PathBuf,
-    /// Manifest hash
-    manifest_hash: Option<FileHash>,
-    /// A combined hash for manifest files of the dependencies
-    deps_hash: String,
-    /// Information about cached dependencies
-    cached_deps: Option<AnalyzedPkgInfo>,
-    /// Compiled user program
-    program: CompiledProgram,
-    /// Maped files
-    mapped_files: MappedFiles,
-    /// Edition of the compiler
-    edition: Option<Edition>,
-    /// Compiler info
-    compiler_info: Option<CompilerInfo>,
-}
-
-/// Data used during symbols computation
-#[derive(Clone)]
-pub struct SymbolsComputationData {
-    /// Outermost definitions in a module (structs, consts, functions), keyed on a ModuleIdent
-    /// string
-    mod_outer_defs: BTreeMap<String, ModuleDefs>,
-    /// A UseDefMap for a given module (needs to be appropriately set before the module
-    /// processing starts) keyed on a ModuleIdent string
-    mod_use_defs: BTreeMap<String, UseDefMap>,
-    /// Uses (references) for a definition at a given location
-    references: BTreeMap<Loc, BTreeSet<UseLoc>>,
-    /// Additional information about a definitions at a given location
-    def_info: BTreeMap<Loc, DefInfo>,
-    /// Module name lengths in access paths for a given module (needs to be appropriately
-    /// set before the module processing starts) keyed on a ModuleIdent string
-    mod_to_alias_lengths: BTreeMap<String, BTreeMap<Position, usize>>,
-}
-
-impl Default for SymbolsComputationData {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl SymbolsComputationData {
-    pub fn new() -> Self {
-        Self {
-            mod_outer_defs: BTreeMap::new(),
-            mod_use_defs: BTreeMap::new(),
-            references: BTreeMap::new(),
-            def_info: BTreeMap::new(),
-            mod_to_alias_lengths: BTreeMap::new(),
-        }
-    }
-}
-
-/// Precomputed information about the package and its dependencies
-/// cached with the purpose of being re-used during the analysis.
-#[derive(Clone)]
-pub struct PrecomputedPkgInfo {
-    /// Hash of the manifest file for a given package
-    manifest_hash: Option<FileHash>,
-    /// Hash of dependency source files
-    deps_hash: String,
-    /// Precompiled deps
-    deps: Arc<FullyCompiledProgram>,
-    /// Symbols computation data
-    deps_symbols_data: Arc<SymbolsComputationData>,
-    /// Compiled user program
-    program: Arc<CompiledProgram>,
-    /// Mapping from file paths to file hashes
-    file_hashes: Arc<BTreeMap<PathBuf, FileHash>>,
-    /// Edition of the compiler used to build this package
-    edition: Option<Edition>,
-    /// Compiler info
-    compiler_info: Option<CompilerInfo>,
-}
-
-/// Location of a use's identifier
-#[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Copy)]
-pub struct UseLoc {
-    /// File where this use identifier starts
-    fhash: FileHash,
-    /// Location where this use identifier starts
-    start: Position,
-    /// Column (on the same line as start)  where this use identifier ends
-    col_end: u32,
-}
-
-/// Type of a function
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub enum FunType {
-    Macro,
-    Entry,
-    Regular,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct VariantInfo {
-    pub name: Name,
-    pub empty: bool,
-    pub positional: bool,
-}
-/// Information about a definition of some identifier
-#[derive(Debug, Clone, Eq, PartialEq)]
-#[allow(clippy::large_enum_variant)]
-pub enum DefInfo {
-    /// Type of an identifier
-    Type(Type),
-    Function(
-        /// Defining module
-        ModuleIdent_,
-        /// Visibility
-        Visibility,
-        /// For example, a macro or entry function
-        FunType,
-        /// Name
-        Symbol,
-        /// Type args
-        Vec<Type>,
-        /// Arg names
-        Vec<Name>,
-        /// Arg types
-        Vec<Type>,
-        /// Ret type
-        Type,
-        /// Doc string
-        Option<String>,
-    ),
-    Struct(
-        /// Defining module
-        ModuleIdent_,
-        /// Name
-        Symbol,
-        /// Visibility
-        Visibility,
-        /// Type args
-        Vec<(Type, bool /* phantom */)>,
-        /// Abilities
-        AbilitySet,
-        /// Field names
-        Vec<Name>,
-        /// Field types
-        Vec<Type>,
-        /// Doc string
-        Option<String>,
-    ),
-    Enum(
-        /// Defining module
-        ModuleIdent_,
-        /// Name
-        Symbol,
-        /// Visibility
-        Visibility,
-        /// Type args
-        Vec<(Type, bool /* phantom */)>,
-        /// Abilities
-        AbilitySet,
-        /// Info about variants
-        Vec<VariantInfo>,
-        /// Doc string
-        Option<String>,
-    ),
-    Variant(
-        /// Defining module of the containing enum
-        ModuleIdent_,
-        /// Name of the containing enum
-        Symbol,
-        /// Variant name
-        Symbol,
-        /// Positional fields?
-        bool,
-        /// Field names
-        Vec<Name>,
-        /// Field types
-        Vec<Type>,
-        /// Doc string
-        Option<String>,
-    ),
-    Field(
-        /// Defining module of the containing struct
-        ModuleIdent_,
-        /// Name of the containing struct
-        Symbol,
-        /// Field name
-        Symbol,
-        /// Field type
-        Type,
-        /// Doc string
-        Option<String>,
-    ),
-    Local(
-        /// Name
-        Symbol,
-        /// Type
-        Type,
-        /// Should displayed definition be preceded by `let`?
-        bool,
-        /// Should displayed definition be preceded by `mut`?
-        bool,
-        /// Location of enum's guard expression (if any) in case
-        /// this local definition represents match pattern's variable
-        Option<Loc>,
-    ),
-    Const(
-        /// Defining module
-        ModuleIdent_,
-        /// Name
-        Symbol,
-        /// Type
-        Type,
-        /// Value
-        Option<String>,
-        /// Doc string
-        Option<String>,
-    ),
-    Module(
-        /// pkg::mod
-        String,
-        /// Doc string
-        Option<String>,
-    ),
-}
-
-/// Information about both the use identifier (source file is specified wherever an instance of this
-/// struct is used) and the definition identifier
-#[derive(Debug, Clone, Eq)]
-pub struct UseDef {
-    /// Column where the (use) identifier location starts on a given line (use this field for
-    /// sorting uses on the line)
-    col_start: u32,
-    /// Column where the (use) identifier location ends on a given line
-    col_end: u32,
-    /// Location of the definition
-    def_loc: Loc,
-    /// Location of the type definition
-    type_def_loc: Option<Loc>,
-}
-
-/// Definition of a struct field
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct FieldDef {
-    pub name: Symbol,
-    pub loc: Loc,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub enum MemberDefInfo {
-    Struct {
-        field_defs: Vec<FieldDef>,
-        positional: bool,
-    },
-    Enum {
-        variants_info: BTreeMap<Symbol, (Loc, Vec<FieldDef>, /* positional */ bool)>,
-    },
-    Fun {
-        attrs: Vec<String>,
-    },
-    Const,
-}
-
-/// Definition of a module member
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
-pub struct MemberDef {
-    pub name_loc: Loc,
-    pub info: MemberDefInfo,
-}
-
-/// Definition of a local (or parameter)
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct LocalDef {
-    /// Location of the definition
-    pub def_loc: Loc,
-    /// Type of definition
-    pub def_type: Type,
-}
-
-impl PartialOrd for LocalDef {
-    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-impl Ord for LocalDef {
-    fn cmp(&self, other: &Self) -> cmp::Ordering {
-        self.def_loc.cmp(&other.def_loc)
-    }
-}
-
-/// Information about call sites relevant to the IDE
-#[derive(Debug, Clone, Ord, PartialOrd, PartialEq, Eq)]
-pub struct CallInfo {
-    /// Is it a dot call?
-    pub dot_call: bool,
-    /// Locations of arguments
-    pub arg_locs: Vec<Loc>,
-    /// Definition of function being called (as an Option as its computed after
-    /// this struct is created)
-    pub def_loc: Option<Loc>,
-}
-
-impl CallInfo {
-    pub fn new(dot_call: bool, args: &[P::Exp]) -> Self {
-        Self {
-            dot_call,
-            arg_locs: args.iter().map(|e| e.loc).collect(),
-            def_loc: None,
-        }
-    }
-}
-
-/// Map from struct name to field order information
-pub type StructFieldOrderInfo = BTreeMap<Symbol, BTreeMap<Symbol, usize>>;
-/// Map from enum name to variant name to field order information
-pub type VariantFieldOrderInfo = BTreeMap<Symbol, BTreeMap<Symbol, BTreeMap<Symbol, usize>>>;
-
-/// Information about field order in structs and enums needed for auto-completion
-/// to be consistent with field order in the source code
-#[derive(Debug, Clone, Ord, PartialOrd, PartialEq, Eq)]
-pub struct FieldOrderInfo {
-    structs: BTreeMap<String, StructFieldOrderInfo>,
-    variants: BTreeMap<String, VariantFieldOrderInfo>,
-}
-
-impl Default for FieldOrderInfo {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl FieldOrderInfo {
-    pub fn new() -> Self {
-        Self {
-            structs: BTreeMap::new(),
-            variants: BTreeMap::new(),
-        }
-    }
-}
-
-/// Module-level definitions and other module-related info
-#[derive(Debug, Clone, Copy, Ord, PartialOrd, PartialEq, Eq)]
-pub enum AutoImportInsertionKind {
-    AfterLastImport,
-    BeforeFirstMember, // when no imports exist
-}
-
-/// Information needed for auto-import insertion. We do our best
-/// to make the insertion fit with what's already in the source file.
-/// In particular, if uses are already preasent, we insert the new import
-/// in the following line keeping the tabulation of the previous import.
-/// If no imports are present, we insert the new import before the first
-/// module member (or before its doc comment if it exists), pushing
-/// this member down but keeping its original tabulation.
-#[derive(Debug, Clone, Copy, Ord, PartialOrd, PartialEq, Eq)]
-pub struct AutoImportInsertionInfo {
-    // Kind of auto-import insertion
-    pub kind: AutoImportInsertionKind,
-    // Position in file where insertion should start
-    pub pos: Position,
-    // Tabulation in number of spaces
-    pub tabulation: usize,
-}
-
-/// Module-level definitions and other module-related info
-#[derive(Debug, Clone, Ord, PartialOrd, PartialEq, Eq)]
-pub struct ModuleDefs {
-    /// File where this module is located
-    pub fhash: FileHash,
-    /// Location where this module is located
-    pub name_loc: Loc,
-    /// Module name
-    pub ident: ModuleIdent_,
-    /// Struct definitions
-    pub structs: BTreeMap<Symbol, MemberDef>,
-    /// Enum definitions
-    pub enums: BTreeMap<Symbol, MemberDef>,
-    /// Const definitions
-    pub constants: BTreeMap<Symbol, MemberDef>,
-    /// Function definitions
-    pub functions: BTreeMap<Symbol, MemberDef>,
-    /// Definitions where the type is not explicitly specified
-    /// and should be inserted as an inlay hint
-    pub untyped_defs: BTreeSet<Loc>,
-    /// Information about calls in this module
-    pub call_infos: BTreeMap<Loc, CallInfo>,
-    /// Position where auto-imports should be inserted
-    pub import_insert_info: Option<AutoImportInsertionInfo>,
-    /// Dependencies summary
-    pub neighbors: UniqueMap<ModuleIdent, Neighbor>,
-}
-
-#[derive(Clone, Debug)]
-pub struct CursorContext {
-    /// Set during typing analysis
-    pub module: Option<ModuleIdent>,
-    /// Set during typing analysis
-    pub defn_name: Option<CursorDefinition>,
-    // TODO: consider making this a vector to hold the whole chain upward
-    /// Set during parsing analysis
-    pub position: CursorPosition,
-    /// Location provided for the cursor
-    pub loc: Loc,
-}
-
-#[derive(Clone, Debug, Copy)]
-pub enum ChainCompletionKind {
-    Type,
-    Function,
-    All,
-}
-
-#[derive(Clone, Debug)]
-pub struct ChainInfo {
-    pub chain: P::NameAccessChain,
-    pub kind: ChainCompletionKind,
-    pub inside_use: bool,
-}
-
-impl ChainInfo {
-    pub fn new(chain: P::NameAccessChain, kind: ChainCompletionKind, inside_use: bool) -> Self {
-        Self {
-            chain,
-            kind,
-            inside_use,
-        }
-    }
-}
-
-impl CursorContext {
-    fn new(loc: Loc) -> Self {
-        CursorContext {
-            module: None,
-            defn_name: None,
-            position: CursorPosition::Unknown,
-            loc,
-        }
-    }
-
-    /// Returns access chain for a match pattern, if any
-    fn find_access_chain_in_match_pattern(&self, p: &P::MatchPattern_) -> Option<ChainInfo> {
-        use ChainCompletionKind as CT;
-        use P::MatchPattern_ as MP;
-        match p {
-            MP::PositionalConstructor(chain, _) => {
-                Some(ChainInfo::new(chain.clone(), CT::Type, false))
-            }
-            MP::FieldConstructor(chain, _) => Some(ChainInfo::new(chain.clone(), CT::Type, false)),
-            MP::Name(_, chain) => Some(ChainInfo::new(chain.clone(), CT::All, false)),
-            MP::Literal(_) | MP::Or(..) | MP::At(..) => None,
-        }
-    }
-
-    /// Returns access chain at cursor position (if any) along with the information of what the chain's
-    /// auto-completed target kind should be, and weather it is part of the use statement.
-    pub fn find_access_chain(&self) -> Option<ChainInfo> {
-        use ChainCompletionKind as CT;
-        use CursorPosition as CP;
-        match &self.position {
-            CP::Exp(sp!(_, exp)) => match exp {
-                P::Exp_::Name(chain) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(chain.clone(), CT::All, false));
-                }
-                P::Exp_::Call(chain, _) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(chain.clone(), CT::Function, false));
-                }
-                P::Exp_::Pack(chain, _) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(chain.clone(), CT::Type, false));
-                }
-                _ => (),
-            },
-            CP::Binding(sp!(_, bind)) => match bind {
-                P::Bind_::Unpack(chain, _) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(*(chain.clone()), CT::Type, false));
-                }
-                _ => (),
-            },
-            CP::Type(sp!(_, ty)) => match ty {
-                P::Type_::Apply(chain) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(*(chain.clone()), CT::Type, false));
-                }
-                _ => (),
-            },
-            CP::Attribute(attr_val) => match &attr_val.value {
-                P::AttributeValue_::ModuleAccess(chain) if chain.loc.contains(&self.loc) => {
-                    return Some(ChainInfo::new(chain.clone(), CT::All, false));
-                }
-                _ => (),
-            },
-            CP::Use(sp!(_, P::Use::Fun { function, ty, .. })) => {
-                if function.loc.contains(&self.loc) {
-                    return Some(ChainInfo::new(*(function.clone()), CT::Function, true));
-                }
-                if ty.loc.contains(&self.loc) {
-                    return Some(ChainInfo::new(*(ty.clone()), CT::Type, true));
-                }
-            }
-            CP::MatchPattern(sp!(_, p)) => return self.find_access_chain_in_match_pattern(p),
-            _ => (),
-        };
-        None
-    }
-
-    /// Returns use declaration at cursor position (if any).
-    pub fn find_use_decl(&self) -> Option<P::Use> {
-        if let CursorPosition::Use(use_) = &self.position {
-            return Some(use_.value.clone());
-        }
-        None
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum CursorPosition {
-    Exp(P::Exp),
-    SeqItem(P::SequenceItem),
-    Binding(P::Bind),
-    Type(P::Type),
-    FieldDefn(P::Field),
-    Parameter(P::Var),
-    DefName,
-    Attribute(P::AttributeValue),
-    Use(Spanned<P::Use>),
-    MatchPattern(P::MatchPattern),
-    Unknown,
-    // FIXME: These two are currently unused because these forms don't have enough location
-    // recorded on them during parsing.
-    DatatypeTypeParameter(P::DatatypeTypeParameter),
-    FunctionTypeParameter((Name, Vec<P::Ability>)),
-}
-
-#[derive(Clone, Debug)]
-pub enum CursorDefinition {
-    Function(P::FunctionName),
-    Constant(P::ConstantName),
-    Struct(P::DatatypeName),
-    Enum(P::DatatypeName),
-}
-
-type LineOffset = u32;
-
-/// Maps a line number to a list of use-def-s on a given line (use-def set is sorted by col_start)
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct UseDefMap(BTreeMap<LineOffset, BTreeSet<UseDef>>);
-
-pub type References = BTreeMap<Loc, BTreeSet<UseLoc>>;
-pub type DefMap = BTreeMap<Loc, DefInfo>;
-pub type FileUseDefs = BTreeMap<PathBuf, UseDefMap>;
-pub type FileModules = BTreeMap<PathBuf, BTreeSet<ModuleDefs>>;
-
-/// Result of the symbolication process
-#[derive(Debug, Clone)]
-pub struct Symbols {
-    /// A map from def locations to all the references (uses)
-    pub references: References,
-    /// A mapping from uses to definitions in a file
-    pub file_use_defs: FileUseDefs,
-    /// A mapping from filePath to ModuleDefs
-    pub file_mods: FileModules,
-    /// Mapped file information for translating locations into positions
-    pub files: MappedFiles,
-    /// Additional information about definitions
-    pub def_info: DefMap,
-    /// IDE Annotation Information from the Compiler
-    pub compiler_info: CompilerInfo,
-    /// Cursor information gathered up during analysis
-    pub cursor_context: Option<CursorContext>,
-}
 
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd)]
-enum RunnerState {
+pub enum RunnerState {
     Run(BTreeSet<PathBuf>),
     Wait,
     Quit,
@@ -725,725 +131,6 @@ enum RunnerState {
 /// Data used during symbolication running and symbolication info updating
 pub struct SymbolicatorRunner {
     mtx_cvar: Arc<(Mutex<RunnerState>, Condvar)>,
-}
-
-impl ModuleDefs {
-    pub fn functions(&self) -> &BTreeMap<Symbol, MemberDef> {
-        &self.functions
-    }
-
-    pub fn structs(&self) -> &BTreeMap<Symbol, MemberDef> {
-        &self.structs
-    }
-
-    pub fn fhash(&self) -> FileHash {
-        self.fhash
-    }
-
-    pub fn untyped_defs(&self) -> &BTreeSet<Loc> {
-        &self.untyped_defs
-    }
-
-    pub fn ident(&self) -> &ModuleIdent_ {
-        &self.ident
-    }
-}
-
-impl fmt::Display for DefInfo {
-    fn fmt(&self, f: &mut fmt::Formatter) -> std::fmt::Result {
-        match self {
-            Self::Type(t) => {
-                // Technically, we could use error_format function here to display the "regular"
-                // type, but the original intent of this function is subtly different that we need
-                // (i.e., to be used by compiler error messages) which, for example, results in
-                // verbosity that is not needed here.
-                //
-                // It also seems like a reasonable idea to be able to tune user experience in the
-                // IDE independently on how compiler error messages are generated.
-                write!(f, "{}", type_to_ide_string(t, /* verbose */ true))
-            }
-            Self::Function(
-                mod_ident,
-                visibility,
-                fun_type,
-                name,
-                type_args,
-                arg_names,
-                arg_types,
-                ret_type,
-                _,
-            ) => {
-                const SINGLE_LINE_TYPE_ARGS_NUM: usize = 2;
-                // The strategy for displaying function signature is as follows:
-                // - if there are more than SINGLE_LINE_TYPE_ARGS_NUM type args,
-                //   they are displayed on separate lines
-                // - "regular" args are always displayed on separate lines, which
-                //   which is motivated by the fact that datatypes are displayed
-                //   in a fully-qualified form (i.e., with package and module name),
-                //   and that makes the function name already long and (likely)
-                //   the length of each individual type also long (modulo primitive
-                //   types of course, but I think we can live with that)
-                let type_args_str = type_args_to_ide_string(
-                    type_args,
-                    /* separate_lines */ type_args.len() > SINGLE_LINE_TYPE_ARGS_NUM,
-                    /* verbose */ true,
-                );
-                let args_str = typed_id_list_to_ide_string(
-                    arg_names, arg_types, '(', ')', /* separate_lines */ true,
-                    /* verbose */ true,
-                );
-                let ret_type_str = ret_type_to_ide_str(ret_type, /* verbose */ true);
-                write!(
-                    f,
-                    "{}{}fun {}{}{}{}{}",
-                    visibility_to_ide_string(visibility),
-                    fun_type_to_ide_string(fun_type),
-                    mod_ident_to_ide_string(mod_ident, None, true),
-                    name,
-                    type_args_str,
-                    args_str,
-                    ret_type_str,
-                )
-            }
-            Self::Struct(
-                mod_ident,
-                name,
-                visibility,
-                type_args,
-                abilities,
-                field_names,
-                field_types,
-                _,
-            ) => {
-                let type_args_str =
-                    datatype_type_args_to_ide_string(type_args, /* verbose */ true);
-                let abilities_str = abilities_to_ide_string(abilities);
-                if field_names.is_empty() {
-                    write!(
-                        f,
-                        "{}struct {}{}{}{} {{}}",
-                        visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident, Some(name), true),
-                        name,
-                        type_args_str,
-                        abilities_str,
-                    )
-                } else {
-                    write!(
-                        f,
-                        "{}struct {}{}{}{} {}",
-                        visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident, Some(name), true),
-                        name,
-                        type_args_str,
-                        abilities_str,
-                        typed_id_list_to_ide_string(
-                            field_names,
-                            field_types,
-                            '{',
-                            '}',
-                            /* separate_lines */ true,
-                            /* verbose */ true
-                        ),
-                    )
-                }
-            }
-            Self::Enum(mod_ident, name, visibility, type_args, abilities, variants, _) => {
-                let type_args_str =
-                    datatype_type_args_to_ide_string(type_args, /* verbose */ true);
-                let abilities_str = abilities_to_ide_string(abilities);
-                if variants.is_empty() {
-                    write!(
-                        f,
-                        "{}enum {}{}{}{} {{}}",
-                        visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident, Some(name), true),
-                        name,
-                        type_args_str,
-                        abilities_str,
-                    )
-                } else {
-                    write!(
-                        f,
-                        "{}enum {}{}{}{} {{\n{}\n}}",
-                        visibility_to_ide_string(visibility),
-                        mod_ident_to_ide_string(mod_ident, Some(name), true),
-                        name,
-                        type_args_str,
-                        abilities_str,
-                        variant_to_ide_string(variants)
-                    )
-                }
-            }
-            Self::Variant(mod_ident, enum_name, name, positional, field_names, field_types, _) => {
-                if field_types.is_empty() {
-                    write!(
-                        f,
-                        "{}{}::{}",
-                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
-                        enum_name,
-                        name
-                    )
-                } else if *positional {
-                    write!(
-                        f,
-                        "{}{}::{}({})",
-                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
-                        enum_name,
-                        name,
-                        type_list_to_ide_string(
-                            field_types,
-                            /* separate_lines */ false,
-                            /* verbose */ true
-                        )
-                    )
-                } else {
-                    write!(
-                        f,
-                        "{}{}::{}{}",
-                        mod_ident_to_ide_string(mod_ident, Some(enum_name), true),
-                        enum_name,
-                        name,
-                        typed_id_list_to_ide_string(
-                            field_names,
-                            field_types,
-                            '{',
-                            '}',
-                            /* separate_lines */ false,
-                            /* verbose */ true,
-                        ),
-                    )
-                }
-            }
-            Self::Field(mod_ident, struct_name, name, t, _) => {
-                write!(
-                    f,
-                    "{}{}\n{}: {}",
-                    mod_ident_to_ide_string(mod_ident, Some(struct_name), true),
-                    struct_name,
-                    name,
-                    type_to_ide_string(t, /* verbose */ true)
-                )
-            }
-            Self::Local(name, t, is_decl, is_mut, _) => {
-                let mut_str = if *is_mut { "mut " } else { "" };
-                if *is_decl {
-                    write!(
-                        f,
-                        "let {}{}: {}",
-                        mut_str,
-                        name,
-                        type_to_ide_string(t, /* verbose */ true)
-                    )
-                } else {
-                    write!(
-                        f,
-                        "{}{}: {}",
-                        mut_str,
-                        name,
-                        type_to_ide_string(t, /* verbose */ true)
-                    )
-                }
-            }
-            Self::Const(mod_ident, name, t, value, _) => {
-                if let Some(v) = value {
-                    write!(
-                        f,
-                        "const {}::{}: {} = {}",
-                        mod_ident,
-                        name,
-                        type_to_ide_string(t, /* verbose */ true),
-                        v
-                    )
-                } else {
-                    write!(
-                        f,
-                        "const {}::{}: {}",
-                        mod_ident,
-                        name,
-                        type_to_ide_string(t, /* verbose */ true)
-                    )
-                }
-            }
-            Self::Module(mod_ident_str, _) => write!(f, "module {mod_ident_str}"),
-        }
-    }
-}
-
-impl fmt::Display for CursorContext {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let CursorContext {
-            module,
-            defn_name,
-            position,
-            loc: _,
-        } = self;
-        writeln!(f, "cursor info:")?;
-        write!(f, "- module: ")?;
-        match module {
-            Some(mident) => writeln!(f, "{mident}"),
-            None => writeln!(f, "None"),
-        }?;
-        write!(f, "- definition: ")?;
-        match defn_name {
-            Some(defn) => match defn {
-                CursorDefinition::Function(name) => writeln!(f, "function {name}"),
-                CursorDefinition::Constant(name) => writeln!(f, "constant {name}"),
-                CursorDefinition::Struct(name) => writeln!(f, "struct {name}"),
-                CursorDefinition::Enum(name) => writeln!(f, "enum {name}"),
-            },
-            None => writeln!(f, "None"),
-        }?;
-        write!(f, "- position: ")?;
-        match position {
-            CursorPosition::Attribute(value) => {
-                writeln!(f, "attribute value")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::Use(value) => {
-                writeln!(f, "use value")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::DefName => {
-                writeln!(f, "defn name")?;
-            }
-            CursorPosition::Unknown => {
-                writeln!(f, "unknown")?;
-            }
-            CursorPosition::Exp(value) => {
-                writeln!(f, "exp")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::SeqItem(value) => {
-                writeln!(f, "seq item")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::Binding(value) => {
-                writeln!(f, "binder")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::Type(value) => {
-                writeln!(f, "type")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::FieldDefn(value) => {
-                writeln!(f, "field")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::Parameter(value) => {
-                writeln!(f, "parameter")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::MatchPattern(value) => {
-                writeln!(f, "match pattern")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::DatatypeTypeParameter(value) => {
-                writeln!(f, "datatype type param")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-            CursorPosition::FunctionTypeParameter(value) => {
-                writeln!(f, "fun type param")?;
-                writeln!(f, "- value: {:#?}", value)?;
-            }
-        }
-        Ok(())
-    }
-}
-
-fn visibility_to_ide_string(visibility: &Visibility) -> String {
-    let mut visibility_str = "".to_string();
-
-    if visibility != &Visibility::Internal {
-        visibility_str.push_str(format!("{} ", visibility).as_str());
-    }
-    visibility_str
-}
-
-pub fn type_args_to_ide_string(type_args: &[Type], separate_lines: bool, verbose: bool) -> String {
-    let mut type_args_str = "".to_string();
-    if !type_args.is_empty() {
-        type_args_str.push('<');
-        if separate_lines {
-            type_args_str.push('\n');
-        }
-        type_args_str.push_str(&type_list_to_ide_string(type_args, separate_lines, verbose));
-        if separate_lines {
-            type_args_str.push('\n');
-        }
-        type_args_str.push('>');
-    }
-    type_args_str
-}
-
-fn datatype_type_args_to_ide_string(type_args: &[(Type, bool)], verbose: bool) -> String {
-    let mut type_args_str = "".to_string();
-    if !type_args.is_empty() {
-        type_args_str.push('<');
-        type_args_str.push_str(&datatype_type_list_to_ide_string(type_args, verbose));
-        type_args_str.push('>');
-    }
-    type_args_str
-}
-
-fn typed_id_list_to_ide_string(
-    names: &[Name],
-    types: &[Type],
-    list_start: char,
-    list_end: char,
-    separate_lines: bool,
-    verbose: bool,
-) -> String {
-    let list = names
-        .iter()
-        .zip(types.iter())
-        .map(|(n, t)| {
-            if separate_lines {
-                format!("\t{}: {}", n.value, type_to_ide_string(t, verbose))
-            } else {
-                format!("{}: {}", n.value, type_to_ide_string(t, verbose))
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(if separate_lines { ",\n" } else { ", " });
-    if separate_lines && !list.is_empty() {
-        format!("{}\n{}\n{}", list_start, list, list_end)
-    } else {
-        format!("{}{}{}", list_start, list, list_end)
-    }
-}
-
-pub fn type_to_ide_string(sp!(_, t): &Type, verbose: bool) -> String {
-    match t {
-        Type_::Unit => "()".to_string(),
-        Type_::Ref(m, r) => format!(
-            "&{}{}",
-            if *m { "mut " } else { "" },
-            type_to_ide_string(r, verbose)
-        ),
-        Type_::Param(tp) => {
-            format!("{}", tp.user_specified_name)
-        }
-        Type_::Apply(_, sp!(_, type_name), ss) => match type_name {
-            TypeName_::Multiple(_) => {
-                format!(
-                    "({})",
-                    type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
-                )
-            }
-            TypeName_::Builtin(name) => {
-                if ss.is_empty() {
-                    format!("{}", name)
-                } else {
-                    format!(
-                        "{}<{}>",
-                        name,
-                        type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
-                    )
-                }
-            }
-            TypeName_::ModuleType(sp!(_, mod_ident), datatype_name) => {
-                let type_args = if ss.is_empty() {
-                    "".to_string()
-                } else {
-                    format!(
-                        "<{}>",
-                        type_list_to_ide_string(ss, /* separate_lines */ false, verbose)
-                    )
-                };
-                if verbose {
-                    format!(
-                        "{}{}{}",
-                        mod_ident_to_ide_string(mod_ident, Some(&datatype_name.value()), true),
-                        datatype_name,
-                        type_args
-                    )
-                } else {
-                    datatype_name.to_string()
-                }
-            }
-        },
-        Type_::Fun(args, ret) => {
-            format!(
-                "|{}| -> {}",
-                type_list_to_ide_string(args, /* separate_lines */ false, verbose),
-                type_to_ide_string(ret, verbose)
-            )
-        }
-        Type_::Anything => "_".to_string(),
-        Type_::Var(_) => "invalid type (var)".to_string(),
-        Type_::UnresolvedError => "unknown type (unresolved)".to_string(),
-    }
-}
-
-pub fn type_list_to_ide_string(types: &[Type], separate_lines: bool, verbose: bool) -> String {
-    types
-        .iter()
-        .map(|t| {
-            if separate_lines {
-                format!("\t{}", type_to_ide_string(t, verbose))
-            } else {
-                type_to_ide_string(t, verbose)
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(if separate_lines { ",\n" } else { ", " })
-}
-
-fn datatype_type_list_to_ide_string(types: &[(Type, bool)], verbose: bool) -> String {
-    types
-        .iter()
-        .map(|(t, phantom)| {
-            if *phantom {
-                format!("phantom {}", type_to_ide_string(t, verbose))
-            } else {
-                type_to_ide_string(t, verbose)
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-pub fn ret_type_to_ide_str(ret_type: &Type, verbose: bool) -> String {
-    match ret_type {
-        sp!(_, Type_::Unit) => "".to_string(),
-        _ => format!(": {}", type_to_ide_string(ret_type, verbose)),
-    }
-}
-/// Conversions of constant values to strings is currently best-effort which is why this function
-/// returns an Option (in the worst case we will display constant name and type but no value).
-fn const_val_to_ide_string(exp: &Exp) -> Option<String> {
-    ast_exp_to_ide_string(exp)
-}
-
-fn ast_exp_to_ide_string(exp: &Exp) -> Option<String> {
-    use UnannotatedExp_ as UE;
-    let sp!(_, e) = &exp.exp;
-    match e {
-        UE::Constant(mod_ident, name) => Some(format!("{mod_ident}::{name}")),
-        UE::Value(v) => Some(ast_value_to_ide_string(v)),
-        UE::Vector(_, _, _, exp) => ast_exp_to_ide_string(exp).map(|s| format!("[{s}]")),
-        UE::Block((_, seq)) | UE::NamedBlock(_, (_, seq)) => {
-            let seq_items = seq
-                .iter()
-                .map(ast_seq_item_to_ide_string)
-                .collect::<Vec<_>>();
-            if seq_items.iter().any(|o| o.is_none()) {
-                // even if only one element cannot be turned into string, don't try displaying block content at all
-                return None;
-            }
-            Some(
-                seq_items
-                    .into_iter()
-                    .map(|o| o.unwrap())
-                    .collect::<Vec<_>>()
-                    .join(", "),
-            )
-        }
-        UE::ExpList(list) => {
-            let items = list
-                .iter()
-                .map(|i| match i {
-                    ExpListItem::Single(exp, _) => ast_exp_to_ide_string(exp),
-                    ExpListItem::Splat(_, exp, _) => ast_exp_to_ide_string(exp),
-                })
-                .collect::<Vec<_>>();
-            if items.iter().any(|o| o.is_none()) {
-                // even if only one element cannot be turned into string, don't try displaying expression list at all
-                return None;
-            }
-            Some(
-                items
-                    .into_iter()
-                    .map(|o| o.unwrap())
-                    .collect::<Vec<_>>()
-                    .join(", "),
-            )
-        }
-        UE::UnaryExp(op, exp) => ast_exp_to_ide_string(exp).map(|s| format!("{op}{s}")),
-
-        UE::BinopExp(lexp, op, _, rexp) => {
-            let Some(ls) = ast_exp_to_ide_string(lexp) else {
-                return None;
-            };
-            let Some(rs) = ast_exp_to_ide_string(rexp) else {
-                return None;
-            };
-            Some(format!("{ls} {op} {rs}"))
-        }
-        _ => None,
-    }
-}
-
-fn ast_seq_item_to_ide_string(sp!(_, seq_item): &SequenceItem) -> Option<String> {
-    use SequenceItem_ as SI;
-    match seq_item {
-        SI::Seq(exp) => ast_exp_to_ide_string(exp),
-        _ => None,
-    }
-}
-
-fn ast_value_to_ide_string(sp!(_, val): &Value) -> String {
-    use Value_ as V;
-    match val {
-        V::Address(addr) => format!("@{}", addr),
-        V::InferredNum(u) => format!("{}", u),
-        V::U8(u) => format!("{}", u),
-        V::U16(u) => format!("{}", u),
-        V::U32(u) => format!("{}", u),
-        V::U64(u) => format!("{}", u),
-        V::U128(u) => format!("{}", u),
-        V::U256(u) => format!("{}", u),
-        V::Bool(b) => format!("{}", b),
-        V::Bytearray(vec) => format!(
-            "[{}]",
-            vec.iter()
-                .map(|v| format!("{}", v))
-                .collect::<Vec<_>>()
-                .join(", ")
-        ),
-    }
-}
-
-/// Creates a string representing a module ID, either on it's owne as in `pkg::module`
-/// or as part of a datatype or function type, in which it should be `pkg::module::`.
-/// If it's part of the datatype, name of the datatype is passed in `datatype_name_opt`.
-pub fn mod_ident_to_ide_string(
-    mod_ident: &ModuleIdent_,
-    datatype_name_opt: Option<&Symbol>,
-    is_access_chain_prefix: bool, // part of access chaing that should end with `::`
-) -> String {
-    use E::Address as A;
-    // the module ID is to be a prefix to a data
-    let suffix = if is_access_chain_prefix { "::" } else { "" };
-    match mod_ident.address {
-        A::Numerical { name, value, .. } => {
-            fn strip_prefix(
-                addr_name: Option<Name>,
-                addr_value: Spanned<NumericalAddress>,
-                mod_ident: &ModuleIdent_,
-                datatype_name_opt: Option<&Symbol>,
-                suffix: &str,
-                pkg_addr: &str,
-                implicit_modules: &[Symbol],
-                implicit_members: &[(Symbol, Symbol, ModuleMemberKind)],
-            ) -> (bool, String) {
-                let pkg_name = match addr_name {
-                    Some(n) => n.to_string(),
-                    None => addr_value.to_string(),
-                };
-
-                let Ok(std_lib_pkg_address) = AccountAddress::from_hex_literal(pkg_addr) else {
-                    // getting stdlib address did not work - use the whole thing
-                    return (false, format!("{pkg_name}::{}{}", mod_ident.module, suffix));
-                };
-                if addr_value.value.into_inner() != std_lib_pkg_address {
-                    // it's not a stdlib package - use the whole thing
-                    return (false, format!("{pkg_name}::{}{}", mod_ident.module, suffix));
-                }
-                // try stripping both package and module if this conversion
-                // is for a datatype, oherwise try only stripping package
-                if let Some(datatype_name) = datatype_name_opt {
-                    if implicit_members.iter().any(
-                        |(implicit_mod_name, implicit_datatype_name, _)| {
-                            mod_ident.module.value() == *implicit_mod_name
-                                && datatype_name == implicit_datatype_name
-                        },
-                    ) {
-                        // strip both package and module (whether its meant to be
-                        // part of access chain or not, if there is not module,
-                        // there should be no `::` at the end)
-                        return (true, "".to_string());
-                    }
-                }
-                if implicit_modules
-                    .iter()
-                    .any(|implicit_mod_name| mod_ident.module.value() == *implicit_mod_name)
-                {
-                    // strip package
-                    return (true, format!("{}{}", mod_ident.module.value(), suffix));
-                }
-                // stripping prefix didn't work - use the whole thing
-                (true, format!("{pkg_name}::{}{}", mod_ident.module, suffix))
-            }
-
-            let (strippable, mut res) = strip_prefix(
-                name,
-                value,
-                mod_ident,
-                datatype_name_opt,
-                suffix,
-                STD_LIB_PKG_ADDRESS,
-                IMPLICIT_STD_MODULES,
-                IMPLICIT_STD_MEMBERS,
-            );
-
-            // check for Sui implicits only if the previous call determined
-            // that the prefix was not strippable with respect to implicits
-            // passed as its arguments
-            if !strippable {
-                (_, res) = strip_prefix(
-                    name,
-                    value,
-                    mod_ident,
-                    datatype_name_opt,
-                    suffix,
-                    SUI_LIB_PKG_ADDRESS,
-                    IMPLICIT_SUI_MODULES,
-                    IMPLICIT_SUI_MEMBERS,
-                );
-            }
-            res
-        }
-        A::NamedUnassigned(n) => format!("{n}::{}", mod_ident.module).to_string(),
-    }
-}
-
-fn fun_type_to_ide_string(fun_type: &FunType) -> String {
-    match fun_type {
-        FunType::Entry => "entry ",
-        FunType::Macro => "macro ",
-        FunType::Regular => "",
-    }
-    .to_string()
-}
-
-fn abilities_to_ide_string(abilities: &AbilitySet) -> String {
-    if abilities.is_empty() {
-        "".to_string()
-    } else {
-        format!(
-            " has {}",
-            abilities
-                .iter()
-                .map(|a| format!("{a}"))
-                .collect::<Vec<_>>()
-                .join(", ")
-        )
-    }
-}
-
-fn variant_to_ide_string(variants: &[VariantInfo]) -> String {
-    // how many variant lines (including optional ellipsis if there
-    // are too many of them) are printed
-    const NUM_PRINTED: usize = 7;
-    let mut vstrings = variants
-        .iter()
-        .enumerate()
-        .map(|(idx, info)| {
-            if idx >= NUM_PRINTED - 1 {
-                "\t/* ... */".to_string()
-            } else if info.empty {
-                format!("\t{}", info.name)
-            } else if info.positional {
-                format!("\t{}( /* ... */ )", info.name)
-            } else {
-                format!("\t{}{{ /* ... */ }}", info.name)
-            }
-        })
-        .collect::<Vec<_>>();
-    vstrings.truncate(NUM_PRINTED);
-    vstrings.join(",\n")
 }
 
 impl SymbolicatorRunner {
@@ -1622,221 +309,6 @@ impl SymbolicatorRunner {
             current_path_opt = current_path.parent();
         }
         None
-    }
-}
-
-impl UseDef {
-    pub fn new(
-        references: &mut References,
-        alias_lengths: &BTreeMap<Position, usize>,
-        use_fhash: FileHash,
-        use_start: Position,
-        def_loc: Loc,
-        use_name: &Symbol,
-        type_def_loc: Option<Loc>,
-    ) -> Self {
-        // Normally, we compute the length of the identifier as the length
-        // of the string that represents it as this string is the same
-        // in the source file and in the AST. However, for aliased module
-        // accesses, the string in the source represents the alias and
-        // the string in the AST represents the actual (non-aliased) module
-        // name - we need to retrieve the correct source-level length
-        // from the map, otherwise on-hover may not work correctly
-        // if AST-level and source-level lengths are different.
-        //
-        // To illustrate it with an example, in the source we may have:
-        //
-        // module Symbols::M9 {
-        //     use Symbols::M1 as ALIAS_M1;
-        //
-        //    struct SomeStruct  {
-        //        some_field: ALIAS_M1::AnotherStruct,
-        //    }
-        // }
-        //
-        // In the (typed) AST we will however have:
-        //
-        // module Symbols::M9 {
-        //     use Symbols::M1 as ALIAS_M1;
-        //
-        //    struct SomeStruct  {
-        //        some_field: M1::AnotherStruct,
-        //    }
-        // }
-        //
-        // As a result, when trying to connect the "use" of module alias with
-        // the module definition, at the level of (typed) AST we will have
-        // identifier of the wrong length which may mess up on-hover and go-to-default
-        // (hovering over a portion of a longer alias may not trigger either).
-
-        let use_name_len = match alias_lengths.get(&use_start) {
-            Some(l) => *l,
-            None => use_name.len(),
-        };
-        let col_end = use_start.character + use_name_len as u32;
-        let use_loc = UseLoc {
-            fhash: use_fhash,
-            start: use_start,
-            col_end,
-        };
-
-        references.entry(def_loc).or_default().insert(use_loc);
-        Self {
-            col_start: use_start.character,
-            col_end,
-            def_loc,
-            type_def_loc,
-        }
-    }
-
-    /// Given a UseDef, modify just the use name and location (to make it represent an alias).
-    pub fn rename_use(
-        &mut self,
-        references: &mut References,
-        new_name: Symbol,
-        new_start: Position,
-        new_fhash: FileHash,
-    ) {
-        self.col_start = new_start.character;
-        self.col_end = new_start.character + new_name.len() as u32;
-        let new_use_loc = UseLoc {
-            fhash: new_fhash,
-            start: new_start,
-            col_end: self.col_end,
-        };
-
-        references
-            .entry(self.def_loc)
-            .or_default()
-            .insert(new_use_loc);
-    }
-
-    pub fn col_start(&self) -> u32 {
-        self.col_start
-    }
-
-    pub fn col_end(&self) -> u32 {
-        self.col_end
-    }
-
-    pub fn def_loc(&self) -> Loc {
-        self.def_loc
-    }
-
-    // use_line is zero-indexed
-    pub fn render(
-        &self,
-        f: &mut dyn std::io::Write,
-        symbols: &Symbols,
-        use_line: u32,
-        use_file_content: &str,
-        def_file_content: &str,
-    ) -> std::io::Result<()> {
-        let UseDef {
-            col_start,
-            col_end,
-            def_loc,
-            type_def_loc,
-        } = self;
-        let uident = use_ident(use_file_content, use_line, *col_start, *col_end);
-        writeln!(f, "Use: '{uident}', start: {col_start}, end: {col_end}")?;
-        let dstart = symbols.files.start_position(def_loc);
-        let dline = dstart.line_offset() as u32;
-        let dcharacter = dstart.column_offset() as u32;
-        let dident = def_ident(def_file_content, dline, dcharacter);
-        writeln!(f, "Def: '{dident}', line: {dline}, def char: {dcharacter}")?;
-        if let Some(ty_loc) = type_def_loc {
-            let tdstart = symbols.files.start_position(ty_loc);
-            let tdline = tdstart.line_offset() as u32;
-            let tdcharacter = tdstart.column_offset() as u32;
-            if let Some((_, type_def_file_content)) = symbols.files.get(&ty_loc.file_hash()) {
-                let type_dident = def_ident(&type_def_file_content, tdline, tdcharacter);
-                writeln!(
-                    f,
-                    "TypeDef: '{type_dident}', line: {tdline}, char: {tdcharacter}"
-                )
-            } else {
-                writeln!(f, "TypeDef: INCORRECT INFO")
-            }
-        } else {
-            writeln!(f, "TypeDef: no info")
-        }
-    }
-}
-
-fn use_ident(use_file_content: &str, use_line: u32, col_start: u32, col_end: u32) -> String {
-    if let Some(line) = use_file_content.lines().nth(use_line as usize) {
-        if let Some((start, _)) = line.char_indices().nth(col_start as usize) {
-            if let Some((end, _)) = line.char_indices().nth(col_end as usize) {
-                return line[start..end].into();
-            }
-        }
-    }
-    "INVALID USE IDENT".to_string()
-}
-
-fn def_ident(def_file_content: &str, def_line: u32, col_start: u32) -> String {
-    if let Some(line) = def_file_content.lines().nth(def_line as usize) {
-        if let Some((start, _)) = line.char_indices().nth(col_start as usize) {
-            let end = line[start..]
-                .char_indices()
-                .find(|(_, c)| !c.is_alphanumeric() && *c != '_' && *c != '$')
-                .map_or(line.len(), |(i, _)| start + i);
-            return line[start..end].into();
-        }
-    }
-    "INVALID DEF IDENT".to_string()
-}
-
-impl Ord for UseDef {
-    fn cmp(&self, other: &Self) -> cmp::Ordering {
-        self.col_start.cmp(&other.col_start)
-    }
-}
-
-impl PartialOrd for UseDef {
-    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl PartialEq for UseDef {
-    fn eq(&self, other: &Self) -> bool {
-        self.col_start == other.col_start
-    }
-}
-
-impl Default for UseDefMap {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl UseDefMap {
-    pub fn new() -> Self {
-        Self(BTreeMap::new())
-    }
-
-    pub fn insert(&mut self, key: u32, val: UseDef) {
-        self.0.entry(key).or_default().insert(val);
-    }
-
-    pub fn get(&self, key: u32) -> Option<BTreeSet<UseDef>> {
-        self.0.get(&key).cloned()
-    }
-
-    pub fn elements(self) -> BTreeMap<u32, BTreeSet<UseDef>> {
-        self.0
-    }
-
-    pub fn count(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn extend(&mut self, use_defs: BTreeMap<u32, BTreeSet<UseDef>>) {
-        for (k, v) in use_defs {
-            self.0.entry(k).or_default().extend(v);
-        }
     }
 }
